+++ app/scripts/VerticalTrack.jsx
@@ -12,6 +12,10 @@ class VerticalTrack extends MoveableTrack {
   getControls(isVisible) {
     return (
       <TrackControl
+        // Whether the button to expand or collapse the track is available
+        // for use. Only available in horizontal and vertical tracks and not
+        // center.
+        expandCollapseAvailable={true}
         imgStyleAdd={STYLES}
         imgStyleClose={STYLES}
         imgStyleMove={STYLES}